requirements.txt
Flask==2.3.3
python-docx==1.1.0
wsgidav==4.4.0
cheroot==10.0.0
